feat: add async support to video segmenter

+++ bot.py
@@ -14,6 +14,7 @@ from disnake.ext import commands, tasks
 
 from NsfwLiveCam import NsfwLiveCam
 from video_segmenter import segment
+from video_segmenter_v2 import VidSegmenter
 
 bot = commands.InteractionBot(intents=disnake.Intents.all())
 logger = logging.getLogger(__name__)
@@ -73,10 +74,11 @@ class Adownloader:
         return dir
 
 
-async def upload(inter: disnake.Interaction, dir: Path):
+async def upload_v2(inter: disnake.Interaction, dir: Path) -> None:
+    segmenter = VidSegmenter()
     if dir.is_file():
         try:
-            dir = segment(dir, 24, Path("."))
+            dir = await segmenter.segment(dir, 24, Path("."))
             await upload(inter, dir)
         except ValueError:
             await inter.send(file=disnake.File(dir))
@@ -86,6 +88,43 @@ async def upload(inter: disnake.Interaction, dir: Path):
                 delete_after=5,
                 allowed_mentions=disnake.AllowedMentions(),
             )
+            dir.unlink()
+            return
+    dir_iter = {x for x in map(lambda x: Path(x), dir.iterdir()) if x.is_file()}
+    to_segment = {file for file in dir_iter if file.stat().st_size / 1024**2 > 24}
+    if to_segment:
+        logger.info(f"{len(to_segment)} files found which are more than 25mb detected")
+        for file in to_segment:
+            try:
+                seg_dir = await segmenter.segment(media=file, max_size=24, save_dir=dir)
+            except:
+                continue
+            file.unlink()
+            await upload(inter, seg_dir)
+    dir_iter = dir_iter - to_segment
+    total_file = [file for file in map(lambda x: disnake.File(x), dir_iter)]
+    file_grps = [total_file[i : i + 10] for i in range(0, len(total_file), 10)]
+    for file_grp in file_grps:
+        await inter.channel.send(files=file_grp)
+    for file in dir_iter:
+        file.unlink()
+    dir.rmdir()
+
+
+async def upload(inter: disnake.Interaction, dir: Path) -> None:
+    if dir.is_file():
+        try:
+            dir = segment(dir, 24, Path("."))
+            await upload(inter, dir)
+        except ValueError:
+            await inter.send(file=disnake.File(dir))
+        finally:
+            await inter.channel.send(
+                f"{inter.author.mention} upload completes",
+                delete_after=5,
+                allowed_mentions=disnake.AllowedMentions(),
+            )
+            dir.unlink()
             return
     dir_iter = {x for x in map(lambda x: Path(x), dir.iterdir()) if x.is_file()}
     to_segment = {file for file in dir_iter if file.stat().st_size / 1024**2 > 24}
@@ -181,10 +220,12 @@ class RecorderView(disnake.ui.View):
 
     @disnake.ui.button(label="Stop Recording", style=disnake.ButtonStyle.red)
     async def red(self, button: disnake.ui.Button, inter: disnake.MessageInteraction):
+        if inter.author.id != inter.message.interaction.author.id:  # type: ignore
+            await inter.send("Recording wasn't started by you!", ephemeral=True)
+            return
+        await inter.response.defer()
         self.process.terminate()
-        await asyncio.sleep(1)
-        await inter.send("Recording Stopped", ephemeral=True)
-        await upload(inter, self.recorder.out_path)
+        await inter.delete_original_response()
 
 
 @bot.slash_command(name="record")
@@ -196,6 +237,13 @@ async def record(inter: disnake.GuildCommandInteraction, model: str):
     await inter.send(
         await recorder.get_thumbnail(), view=RecorderView(process, recorder)
     )
+    await process.wait()
+    try:
+        await upload_v2(inter, recorder.out_path)
+    except FileNotFoundError:
+        await inter.edit_original_response(
+            "Model Is Currenlty Offline or in Private Show"
+        )
 
 
 if __name__ == "__main__":
